debian/carpc/opt/carpc/backend/requirements.txt
Django==1.6.4
#django-bower==5.0.1
django-supervisor==0.3.2
picamera==1.4
pyserial==2.7
South==0.8.4
requests==2.3.0
BeautifulSoup==3.2.1
PyCrypto==2.6.1
pytz==2014.4